Introducing a further abstraction level
