+++ src/main/java/edu/ie3/datamodel/models/UniqueEntity.java
@@ -10,7 +10,10 @@ import java.util.UUID;
 
 /** Supplies every subclass with a generated UUID, making it unique */
 public abstract class UniqueEntity {
-  private final UUID uuid;
+  /** Field name of {@link UniqueEntity} uuid */
+  public static final String UUID_FIELD_NAME = "uuid";
+
+  protected final UUID uuid;
 
   public UniqueEntity() {
     uuid = UUID.randomUUID();
